Remove debug and unused files, simplify theme system
