--- css-debug.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>CSS Variables Test</title>
-    <style>
-        :root {
-            --test-color: red;
-            --test-bg: blue;
-        }
-        
-        body[data-theme="dark"] {
-            --test-color: white;
-            --test-bg: black;
-        }
-        
-        body[data-theme="light"] {
-            --test-color: black;
-            --test-bg: white;
-        }
-        
-        .test-div {
-            background-color: var(--test-bg);
-            color: var(--test-color);
-            padding: 20px;
-            margin: 20px;
-            border: 2px solid var(--test-color);
-        }
-    </style>
-</head>
-<body data-theme="light">
-    <h1>CSS Variables Test</h1>
-    
-    <div class="test-div">
-        <p>This div should change colors when you change the theme.</p>
-        <p>Current theme: <span id="currentTheme">light</span></p>
-    </div>
-    
-    <button onclick="changeTheme('dark')">Set Dark Theme</button>
-    <button onclick="changeTheme('light')">Set Light Theme</button>
-    
-    <script>
-        function changeTheme(theme) {
-            document.body.setAttribute('data-theme', theme);
-            document.getElementById('currentTheme').textContent = theme;
-            
-            const computedStyle = getComputedStyle(document.body);
-            console.log('Theme:', theme);
-            console.log('--test-color:', computedStyle.getPropertyValue('--test-color'));
-            console.log('--test-bg:', computedStyle.getPropertyValue('--test-bg'));
-        }
-        
-        // Test on load
-        console.log('Initial CSS variables:');
-        const computedStyle = getComputedStyle(document.body);
-        console.log('--test-color:', computedStyle.getPropertyValue('--test-color'));
-        console.log('--test-bg:', computedStyle.getPropertyValue('--test-bg'));
-    </script>
-</body>
-</html>
