--- css/fix-themes.css
@@ -1,341 +1,0 @@
-/* ===== FIX FOR THEME SYSTEM ===== */
-/* This file fixes the white text on white background issue */
-
-/* Global theme application */
-body {
-    background-color: var(--bg-primary) !important;
-    color: var(--text-primary) !important;
-    transition: background-color 0.3s ease, color 0.3s ease;
-}
-
-/* Force text colors for all elements */
-* {
-    color: inherit;
-}
-
-/* Specific text color fixes */
-h1, h2, h3, h4, h5, h6 {
-    color: var(--text-primary) !important;
-}
-
-p, span, div, a {
-    color: var(--text-primary) !important;
-}
-
-/* Navbar fixes */
-.navbar {
-    background: rgba(255, 255, 255, 0.9) !important;
-    backdrop-filter: blur(20px);
-    border-bottom: 1px solid rgba(255, 255, 255, 0.2);
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
-}
-
-.navbar-brand {
-    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-    font-weight: 800;
-    font-size: 1.5rem;
-}
-
-.nav-link {
-    color: #333 !important;
-}
-
-.nav-link:hover {
-    color: #667eea !important;
-}
-
-/* Search input fix */
-.search-input {
-    background: rgba(255, 255, 255, 0.8) !important;
-    border: 1px solid rgba(255, 255, 255, 0.2) !important;
-    color: #333 !important;
-}
-
-.search-input::placeholder {
-    color: #666 !important;
-}
-
-/* Theme toggle fix */
-.theme-toggle {
-    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%) !important;
-    border: none !important;
-    color: white !important;
-    font-weight: 600;
-    padding: 0.75rem 1.5rem;
-    border-radius: 16px;
-    transition: all 0.3s ease;
-    box-shadow: 0 4px 16px rgba(102, 126, 234, 0.3);
-}
-
-.theme-toggle:hover {
-    transform: translateY(-2px);
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.2);
-}
-
-/* Device selector overlay fix */
-.device-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background: rgba(0, 0, 0, 0.8);
-    backdrop-filter: blur(10px);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    z-index: 9999;
-}
-
-.device-content {
-    background: white;
-    border-radius: 32px;
-    padding: 48px;
-    text-align: center;
-    max-width: 500px;
-    width: 90%;
-    box-shadow: 0 16px 64px rgba(0, 0, 0, 0.2);
-    border: 1px solid #e2e8f0;
-}
-
-.device-header h2 {
-    color: #333 !important;
-    margin-bottom: 8px;
-    font-size: 2rem;
-    font-weight: 800;
-}
-
-.device-header p {
-    color: #666 !important;
-    margin-bottom: 32px;
-    font-size: 1.1rem;
-}
-
-.device-options {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 24px;
-    margin: 32px 0;
-}
-
-.device-option {
-    background: #f8fafc;
-    border: 2px solid #e2e8f0;
-    border-radius: 16px;
-    padding: 24px;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    text-align: center;
-}
-
-.device-option:hover {
-    border-color: #667eea;
-    transform: translateY(-5px);
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.15);
-}
-
-.device-option i {
-    font-size: 2rem;
-    color: #667eea;
-    margin-bottom: 8px;
-    display: block;
-}
-
-.device-option span {
-    color: #333 !important;
-    font-weight: 600;
-    font-size: 0.9rem;
-}
-
-.device-footer {
-    color: #666 !important;
-    font-size: 0.9rem;
-    margin-top: 24px;
-}
-
-/* Hero section fix */
-.hero-section {
-    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-    padding: 48px 0;
-    position: relative;
-    overflow: hidden;
-}
-
-.hero-content {
-    text-align: center;
-    color: white !important;
-    position: relative;
-    z-index: 2;
-}
-
-.hero-title {
-    font-size: 4rem;
-    font-weight: 900;
-    margin-bottom: 24px;
-    color: white !important;
-    text-shadow: 0 0 20px rgba(255, 255, 255, 0.5);
-}
-
-.hero-subtitle {
-    font-size: 1.25rem;
-    margin-bottom: 32px;
-    color: rgba(255, 255, 255, 0.9) !important;
-}
-
-/* Stats container fix */
-.stats-container {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-    gap: 24px;
-    margin: 32px 0;
-}
-
-.stat-card {
-    background: rgba(255, 255, 255, 0.9);
-    backdrop-filter: blur(20px);
-    border: 1px solid rgba(255, 255, 255, 0.2);
-    border-radius: 16px;
-    padding: 24px;
-    text-align: center;
-    transition: all 0.3s ease;
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.08);
-}
-
-.stat-card:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.15);
-}
-
-.stat-number {
-    font-size: 2.5rem;
-    font-weight: 800;
-    color: #667eea;
-    margin-bottom: 8px;
-}
-
-.stat-label {
-    color: #666 !important;
-    font-weight: 600;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-}
-
-/* Trending card fix */
-.trending-card {
-    background: rgba(255, 255, 255, 0.9);
-    backdrop-filter: blur(20px);
-    border: 1px solid rgba(255, 255, 255, 0.2);
-    border-radius: 16px;
-    padding: 24px;
-    margin-bottom: 24px;
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.08);
-    position: relative;
-    overflow: hidden;
-}
-
-.trending-card::before {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    height: 3px;
-    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-    animation: conic-gradient 3s linear infinite;
-}
-
-.trending-header {
-    display: flex;
-    align-items: center;
-    margin-bottom: 16px;
-}
-
-.trending-icon {
-    font-size: 1.5rem;
-    color: #f59e0b;
-    margin-right: 8px;
-    animation: flame 1s ease-in-out infinite alternate;
-}
-
-.trending-title {
-    font-weight: 700;
-    color: #333 !important;
-    font-size: 1.1rem;
-}
-
-/* Button fixes */
-.btn-primary {
-    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%) !important;
-    border: none !important;
-    color: white !important;
-    font-weight: 600;
-    padding: 14px 32px;
-    border-radius: 16px;
-    transition: all 0.3s ease;
-    box-shadow: 0 4px 16px rgba(102, 126, 234, 0.3);
-}
-
-.btn-primary:hover {
-    transform: translateY(-2px);
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.2);
-}
-
-.btn-outline-light {
-    background: transparent !important;
-    border: 2px solid white !important;
-    color: white !important;
-    font-weight: 600;
-    padding: 14px 32px;
-    border-radius: 16px;
-    transition: all 0.3s ease;
-}
-
-.btn-outline-light:hover {
-    background: white !important;
-    color: #667eea !important;
-    transform: translateY(-2px);
-}
-
-/* Animations */
-@keyframes conic-gradient {
-    0% { transform: rotate(0deg); }
-    100% { transform: rotate(360deg); }
-}
-
-@keyframes flame {
-    0% { transform: scale(1) rotate(-5deg); }
-    100% { transform: scale(1.1) rotate(5deg); }
-}
-
-@keyframes glow {
-    0% { filter: drop-shadow(0 0 5px rgba(255, 255, 255, 0.5)); }
-    100% { filter: drop-shadow(0 0 20px rgba(255, 255, 255, 0.8)); }
-}
-
-/* Responsive fixes */
-@media (max-width: 768px) {
-    .hero-title {
-        font-size: 2.5rem;
-    }
-    
-    .stats-container {
-        grid-template-columns: 1fr;
-    }
-    
-    .device-options {
-        grid-template-columns: 1fr;
-    }
-}
-
-@media (max-width: 480px) {
-    .hero-title {
-        font-size: 2rem;
-    }
-    
-    .hero-subtitle {
-        font-size: 1rem;
-    }
-}
