+++ includes/footer.php
@@ -137,7 +137,7 @@
 
 <!-- Custom JavaScript -->
 <script src="js/main.js"></script>
-<script src="js/themes-ultra-premium.js"></script>
+<script src="js/themes.js"></script>
 
 <!-- Initialize Theme Manager -->
 <script>
