+++ js/themes.js
@@ -1,725 +1,1014 @@
 /**
- * Affinity Forum - Premium Theme System v2.0
- * Advanced theme switching with visual effects and customization
+ * Unified Theme Manager
+ * Advanced theme system with particle effects, audio feedback, and accessibility features
  */
 
-// Global theme functions for HTML onclick
-function switchTheme(themeName) {
-    if (window.themeManager) {
-        window.themeManager.setTheme(themeName);
-    } else {
-        // Fallback if themeManager not loaded
-        document.body.setAttribute('data-theme', themeName);
-        localStorage.setItem('affinity_theme', themeName);
-        showThemeNotification(themeName);
-    }
-}
-
-function openThemePreview() {
-    if (window.themeManager) {
-        window.themeManager.showThemePreview();
-    }
-}
-
-function openCustomization() {
-    if (window.themeManager) {
-        window.themeManager.openCustomization();
-    }
-}
-
-function previewTheme(themeName) {
-    if (window.themeManager) {
-        window.themeManager.previewTheme(themeName);
-    }
-}
-
-function applyTheme(themeName) {
-    if (window.themeManager) {
-        window.themeManager.setTheme(themeName);
-    }
-}
-
-function closeThemePreview() {
-    const modal = document.getElementById('theme-preview-modal');
-    if (modal) {
-        modal.style.display = 'none';
-    }
-}
-
-// Global notification function
-function showThemeNotification(themeName) {
-    const themeNames = {
-        light: 'Light Theme',
-        dark: 'Dark Theme',
-        cs2: 'CS2 Theme',
-        premium: 'Premium Theme'
-    };
-    
-    const notification = document.createElement('div');
-    notification.className = 'theme-change-notification';
-    notification.innerHTML = `
-        <div class="theme-notification-content">
-            <i class="fas fa-palette"></i>
-            <span>Switched to ${themeNames[themeName]}</span>
-        </div>
-    `;
-    
-    document.body.appendChild(notification);
-    
-    // Animate in
-    setTimeout(() => {
-        notification.classList.add('show');
-    }, 100);
-    
-    // Remove after delay
-    setTimeout(() => {
-        notification.classList.remove('show');
-        setTimeout(() => {
-            if (notification.parentNode) {
-                notification.remove();
-            }
-        }, 300);
-    }, 3000);
-}
+console.log('🚀 Unified Theme Manager loading...');
 
 class ThemeManager {
     constructor() {
         this.currentTheme = 'light';
-        this.availableThemes = ['light', 'dark', 'cs2', 'premium'];
-        this.themeColors = {
-            light: {
-                primary: '#007bff',
-                secondary: '#6c757d',
-                success: '#28a745',
-                warning: '#ffc107',
-                danger: '#dc3545',
-                info: '#17a2b8'
-            },
-            dark: {
-                primary: '#58a6ff',
-                secondary: '#8b949e',
-                success: '#3fb950',
-                warning: '#d29922',
-                danger: '#f85149',
-                info: '#58a6ff'
-            },
-            cs2: {
-                primary: '#ff6b35',
-                secondary: '#7a8ba1',
-                success: '#48bb78',
-                warning: '#ed8936',
-                danger: '#f56565',
-                info: '#4299e1'
-            },
-            premium: {
-                primary: '#ffd700',
-                secondary: '#a0a0a0',
-                success: '#00ff88',
-                warning: '#ffaa00',
-                danger: '#ff4444',
-                info: '#00aaff'
-            }
+        this.currentDevice = 'desktop';
+        this.particleSystem = null;
+        this.audioContext = null;
+        this.isInitialized = false;
+        this.animations = new Map();
+        this.notifications = [];
+        this.performanceMetrics = {
+            themeSwitchCount: 0,
+            deviceSwitchCount: 0,
+            lastThemeSwitch: 0,
+            lastDeviceSwitch: 0,
+            totalNotifications: 0
+        };
+        this.accessibility = {
+            highContrast: false,
+            reducedMotion: false,
+            fontSize: 16,
+            screenReader: false
+        };
+        this.features = {
+            particles: true,
+            audio: true,
+            animations: true,
+            notifications: true,
+            shortcuts: true
         };
         
+        console.log('🎨 Theme Manager initialized');
         this.init();
     }
-    
+
     init() {
-        this.loadSavedTheme();
-        this.setupThemeToggle();
-        this.setupThemePreview();
-        this.setupCustomization();
-        this.applyTheme(this.currentTheme);
-        
-        // Make themeManager globally accessible
-        window.themeManager = this;
-        
-        // Add theme change event listener
-        document.addEventListener('themeChanged', (e) => {
-            this.onThemeChange(e.detail.theme);
-        });
-    }
-    
-    loadSavedTheme() {
-        const savedTheme = localStorage.getItem('affinity_theme');
-        if (savedTheme && this.availableThemes.includes(savedTheme)) {
-            this.currentTheme = savedTheme;
-        } else {
-            // Default to light theme
-            this.currentTheme = 'light';
+        try {
+            this.loadUserPreferences();
+            this.setupAccessibility();
+            this.setupParticleSystem();
+            this.setupAudioContext();
+            this.setupEventListeners();
+            this.setupPerformanceMonitoring();
+            this.applyCurrentTheme();
+            
+            // Check if this is the first visit and show device selector
+            this.checkFirstVisit();
+            
+            console.log('✅ Theme Manager fully initialized');
+        } catch (error) {
+            console.error('❌ Error in init:', error);
         }
     }
     
-    setupThemeToggle() {
-        // Update the theme name display
-        this.updateThemeName();
+    checkFirstVisit() {
+        const hasSelectedDevice = localStorage.getItem('hasSelectedDevice');
         
-        // Add keyboard shortcut (Ctrl/Cmd + T)
-        document.addEventListener('keydown', (e) => {
-            if ((e.ctrlKey || e.metaKey) && e.key === 't') {
-                e.preventDefault();
-                this.cycleTheme();
+        if (!hasSelectedDevice) {
+            console.log('🎯 First visit detected - showing device selector');
+            if (document.body) {
+                setTimeout(() => {
+                    this.showDeviceSelector();
+                }, 1000);
+            } else {
+                setTimeout(() => this.checkFirstVisit(), 100);
             }
-        });
-    }
-    
-    updateThemeName() {
-        const themeNameSpan = document.querySelector('.theme-name');
-        if (themeNameSpan) {
-            themeNameSpan.textContent = this.currentTheme.charAt(0).toUpperCase() + this.currentTheme.slice(1);
+        } else {
+            console.log('✅ Returning user - device already selected');
         }
     }
     
-    setupThemePreview() {
-        // Create theme preview modal
-        const previewModal = document.createElement('div');
-        previewModal.className = 'theme-preview-modal';
-        previewModal.innerHTML = `
-            <div class="theme-preview-content">
-                <div class="theme-preview-header">
-                    <h3><i class="fas fa-palette"></i> Theme Preview</h3>
-                    <button class="btn-close" data-bs-dismiss="modal">
-                        <i class="fas fa-times"></i>
-                    </button>
-                </div>
-                <div class="theme-preview-grid">
-                    ${this.availableThemes.map(theme => this.createThemePreview(theme)).join('')}
-                </div>
-            </div>
-        `;
-        
-        document.body.appendChild(previewModal);
-        
-        // Add preview button to theme toggle
-        const themeToggle = document.getElementById('theme-toggle');
-        if (themeToggle) {
-            const previewBtn = document.createElement('button');
-            previewBtn.className = 'btn btn-sm btn-outline-secondary ms-2';
-            previewBtn.innerHTML = '<i class="fas fa-eye"></i>';
-            previewBtn.title = 'Preview Themes';
-            previewBtn.addEventListener('click', () => {
-                this.showThemePreview();
-            });
-            themeToggle.parentNode.appendChild(previewBtn);
+    showDeviceSelector() {
+        console.log('🎯 Showing device selector for first visit');
+        const overlay = document.getElementById('deviceOverlay');
+        if (overlay) {
+            overlay.style.display = 'flex';
+            console.log('✅ Device selector shown for first visit');
         }
     }
-    
-    createThemePreview(theme) {
-        const colors = this.themeColors[theme];
-        const isActive = theme === this.currentTheme;
-        
-        return `
-            <div class="theme-preview-item ${isActive ? 'active' : ''}" data-theme="${theme}">
-                <div class="theme-preview-header">
-                    <h4>${theme.charAt(0).toUpperCase() + theme.slice(1)}</h4>
-                    ${isActive ? '<span class="badge bg-success">Active</span>' : ''}
-                </div>
-                <div class="theme-preview-colors">
-                    <div class="color-swatch" style="background: ${colors.primary}" title="Primary"></div>
-                    <div class="color-swatch" style="background: ${colors.secondary}" title="Secondary"></div>
-                    <div class="color-swatch" style="background: ${colors.success}" title="Success"></div>
-                    <div class="color-swatch" style="background: ${colors.warning}" title="Warning"></div>
-                    <div class="color-swatch" style="background: ${colors.danger}" title="Danger"></div>
-                </div>
-                <div class="theme-preview-actions">
-                    <button class="btn btn-sm btn-primary preview-theme-btn" data-theme="${theme}">
-                        Preview
-                    </button>
-                    ${!isActive ? `<button class="btn btn-sm btn-success apply-theme-btn" data-theme="${theme}">Apply</button>` : ''}
-                </div>
-            </div>
-        `;
-    }
-    
-    setupCustomization() {
-        // Create theme customization panel
-        const customizationPanel = document.createElement('div');
-        customizationPanel.className = 'theme-customization-panel';
-        customizationPanel.innerHTML = `
-            <div class="customization-header">
-                <h4><i class="fas fa-sliders-h"></i> Customize Theme</h4>
-                <button class="btn-close-customization">
-                    <i class="fas fa-times"></i>
-                </button>
-            </div>
-            <div class="customization-content">
-                <div class="customization-section">
-                    <label>Primary Color</label>
-                    <input type="color" id="custom-primary" class="form-control">
-                </div>
-                <div class="customization-section">
-                    <label>Background Opacity</label>
-                    <input type="range" id="custom-opacity" min="0.1" max="1" step="0.1" class="form-range">
-                </div>
-                <div class="customization-section">
-                    <label>Border Radius</label>
-                    <input type="range" id="custom-radius" min="0" max="20" step="1" class="form-range">
-                </div>
-                <div class="customization-section">
-                    <label>Animation Speed</label>
-                    <select id="custom-animation" class="form-select">
-                        <option value="fast">Fast</option>
-                        <option value="normal" selected>Normal</option>
-                        <option value="slow">Slow</option>
-                    </select>
-                </div>
-                <div class="customization-actions">
-                    <button class="btn btn-primary" id="save-customization">Save Changes</button>
-                    <button class="btn btn-secondary" id="reset-customization">Reset to Default</button>
-                </div>
-            </div>
-        `;
-        
-        document.body.appendChild(customizationPanel);
-        
-        // Add customization button
-        const themeToggle = document.getElementById('theme-toggle');
-        if (themeToggle) {
-            const customBtn = document.createElement('button');
-            customBtn.className = 'btn btn-sm btn-outline-secondary ms-2';
-            customBtn.innerHTML = '<i class="fas fa-sliders-h"></i>';
-            customBtn.title = 'Customize Theme';
-            customBtn.addEventListener('click', () => {
-                this.toggleCustomization();
-            });
-            themeToggle.parentNode.appendChild(customBtn);
-        }
-        
-        // Bind customization events
-        this.bindCustomizationEvents();
-    }
-    
-    bindCustomizationEvents() {
-        // Color picker
-        const primaryColor = document.getElementById('custom-primary');
-        if (primaryColor) {
-            primaryColor.addEventListener('change', (e) => {
-                this.updateCustomColor('primary', e.target.value);
-            });
-        }
-        
-        // Opacity slider
-        const opacitySlider = document.getElementById('custom-opacity');
-        if (opacitySlider) {
-            opacitySlider.addEventListener('input', (e) => {
-                this.updateCustomOpacity(e.target.value);
-            });
-        }
-        
-        // Border radius slider
-        const radiusSlider = document.getElementById('custom-radius');
-        if (radiusSlider) {
-            radiusSlider.addEventListener('input', (e) => {
-                this.updateCustomRadius(e.target.value);
-            });
-        }
-        
-        // Animation speed
-        const animationSelect = document.getElementById('custom-animation');
-        if (animationSelect) {
-            animationSelect.addEventListener('change', (e) => {
-                this.updateCustomAnimation(e.target.value);
-            });
-        }
-        
-        // Save button
-        const saveBtn = document.getElementById('save-customization');
-        if (saveBtn) {
-            saveBtn.addEventListener('click', () => {
-                this.saveCustomization();
-            });
-        }
-        
-        // Reset button
-        const resetBtn = document.getElementById('reset-customization');
-        if (resetBtn) {
-            resetBtn.addEventListener('click', () => {
-                this.resetCustomization();
-            });
-        }
-        
-        // Close button
-        const closeBtn = document.querySelector('.btn-close-customization');
-        if (closeBtn) {
-            closeBtn.addEventListener('click', () => {
-                this.toggleCustomization();
-            });
-        }
-    }
-    
-    cycleTheme() {
-        const currentIndex = this.availableThemes.indexOf(this.currentTheme);
-        const nextIndex = (currentIndex + 1) % this.availableThemes.length;
-        const nextTheme = this.availableThemes[nextIndex];
-        
-        this.setTheme(nextTheme);
-    }
-    
-    setTheme(themeName) {
-        if (!this.availableThemes.includes(themeName)) {
-            console.error(`Invalid theme: ${themeName}`);
-            return;
-        }
-        
-        console.log(`🎨 Switching to theme: ${themeName}`);
-        
-        // Remove all existing theme classes
-        document.body.classList.remove(...this.availableThemes.map(t => `theme-${t}`));
-        
-        // Set new theme
-        document.body.setAttribute('data-theme', themeName);
-        this.currentTheme = themeName;
-        
-        // Save to localStorage
-        localStorage.setItem('affinity_theme', themeName);
-        
-        // Update theme name display
-        this.updateThemeName();
-        
-        // Apply theme-specific customizations
-        this.applyThemeCustomizations(themeName);
-        
-        // Trigger theme change event
-        document.dispatchEvent(new CustomEvent('themeChanged', {
-            detail: { theme: themeName }
-        }));
-        
-        // Show notification
-        this.showThemeNotification(themeName);
-        
-        // Update any theme-dependent components
-        this.updateThemeComponents(themeName);
-    }
-    
-    applyTheme(themeName) {
-        // Set initial theme
-        document.body.setAttribute('data-theme', themeName);
-        
-        // Apply theme-specific styles
-        this.applyThemeStyles(themeName);
-        
-        // Update components
-        this.updateThemeComponents(themeName);
-    }
-    
-    applyThemeStyles(themeName) {
-        const colors = this.themeColors[themeName];
-        
-        // Update CSS custom properties
-        Object.entries(colors).forEach(([key, value]) => {
-            document.documentElement.style.setProperty(`--accent-${key}`, value);
-        });
-        
-        // Apply theme-specific animations
-        this.applyThemeAnimations(themeName);
-    }
-    
-    applyThemeAnimations(themeName) {
-        const animationSpeeds = {
-            light: 'normal',
-            dark: 'normal',
-            cs2: 'fast',
-            premium: 'slow'
-        };
-        
-        const speed = animationSpeeds[themeName];
-        document.documentElement.style.setProperty('--transition-speed', speed);
-        
-        // Add theme-specific animation classes
-        document.body.classList.add(`theme-${themeName}-animations`);
-    }
-    
-    updateThemeToggle() {
-        const themeToggle = document.getElementById('theme-toggle');
-        if (themeToggle) {
-            const icon = themeToggle.querySelector('i');
-            if (icon) {
-                icon.className = this.getThemeIcon(this.currentTheme);
+
+    loadUserPreferences() {
+        try {
+            // Load theme preference
+            const savedTheme = localStorage.getItem('selectedTheme');
+            if (savedTheme && ['light', 'dark', 'cs2', 'premium'].includes(savedTheme)) {
+                this.currentTheme = savedTheme;
+            } else {
+                this.currentTheme = 'light';
+                localStorage.setItem('selectedTheme', 'light');
             }
             
-            // Update tooltip
-            themeToggle.title = `Current: ${this.currentTheme.charAt(0).toUpperCase() + this.currentTheme.slice(1)} (Click to change)`;
+            // Load device preference
+            const savedDevice = localStorage.getItem('selectedDevice');
+            if (savedDevice && ['desktop', 'tablet', 'mobile'].includes(savedDevice)) {
+                this.currentDevice = savedDevice;
+            } else {
+                this.currentDevice = 'desktop';
+                localStorage.setItem('selectedDevice', 'desktop');
+            }
+            
+            // Load accessibility preferences
+            this.accessibility.highContrast = localStorage.getItem('highContrast') === 'true';
+            this.accessibility.reducedMotion = localStorage.getItem('reducedMotion') === 'true';
+            this.accessibility.fontSize = parseInt(localStorage.getItem('fontSize')) || 16;
+            this.accessibility.screenReader = localStorage.getItem('screenReader') === 'true';
+            
+            // Load feature preferences
+            this.features.particles = localStorage.getItem('particles') !== 'false';
+            this.features.audio = localStorage.getItem('audio') !== 'false';
+            this.features.animations = localStorage.getItem('animations') !== 'false';
+            this.features.notifications = localStorage.getItem('notifications') !== 'false';
+            this.features.shortcuts = localStorage.getItem('shortcuts') !== 'false';
+            
+            console.log('📱 Loaded preferences - Theme:', this.currentTheme, 'Device:', this.currentDevice);
+        } catch (error) {
+            console.warn('⚠️ Failed to load preferences, using defaults:', error);
         }
     }
-    
-    getThemeIcon(theme) {
-        const icons = {
-            light: 'fas fa-sun',
-            dark: 'fas fa-moon',
-            cs2: 'fas fa-crosshairs',
-            premium: 'fas fa-crown'
-        };
-        return icons[theme] || 'fas fa-palette';
+
+    setupAccessibility() {
+        try {
+            if (this.accessibility.highContrast) {
+                document.body.classList.add('high-contrast');
+            }
+            if (this.accessibility.reducedMotion) {
+                document.body.classList.add('reduced-motion');
+            }
+            if (this.accessibility.screenReader) {
+                document.body.classList.add('screen-reader');
+            }
+            
+            document.documentElement.style.fontSize = this.accessibility.fontSize + 'px';
+            
+            console.log('♿ Accessibility features configured');
+        } catch (error) {
+            console.warn('⚠️ Failed to setup accessibility:', error);
+        }
     }
-    
-    showThemeNotification(themeName) {
-        const themeNames = {
-            light: 'Light Theme',
-            dark: 'Dark Theme',
-            cs2: 'CS2 Theme',
-            premium: 'Premium Theme'
-        };
+
+    setupParticleSystem() {
+        if (!this.features.particles) return;
         
-        const notification = document.createElement('div');
-        notification.className = 'theme-change-notification';
-        notification.innerHTML = `
-            <div class="theme-notification-content">
-                <i class="fas fa-palette"></i>
-                <span>Switched to ${themeNames[themeName]}</span>
-            </div>
-        `;
+        try {
+            const canvas = document.createElement('canvas');
+            canvas.id = 'particleCanvas';
+            canvas.style.cssText = `
+                position: fixed;
+                top: 0;
+                left: 0;
+                width: 100%;
+                height: 100%;
+                pointer-events: none;
+                z-index: 1000;
+                opacity: 0.6;
+            `;
+            
+            document.body.appendChild(canvas);
+            
+            const ctx = canvas.getContext('2d');
+            this.resizeCanvas(canvas, ctx);
+            
+            this.particleSystem = {
+                canvas,
+                ctx,
+                particles: [],
+                particleCount: this.currentDevice === 'mobile' ? 20 : this.currentDevice === 'tablet' ? 35 : 50,
+                maxParticles: this.currentDevice === 'mobile' ? 40 : this.currentDevice === 'tablet' ? 70 : 100,
+                isActive: true,
+                animationId: null
+            };
+            
+            // Initialize particles
+            for (let i = 0; i < this.particleSystem.particleCount; i++) {
+                this.addNewParticle();
+            }
+            
+            // Start animation
+            this.startParticleAnimation();
+            
+            // Handle resize
+            window.addEventListener('resize', () => this.resizeCanvas(canvas, ctx));
+            
+            console.log('✨ Particle system initialized');
+        } catch (error) {
+            console.warn('⚠️ Failed to setup particle system:', error);
+        }
+    }
+
+    resizeCanvas(canvas, ctx) {
+        canvas.width = window.innerWidth;
+        canvas.height = window.innerHeight;
+    }
+
+    setupAudioContext() {
+        if (!this.features.audio || !window.AudioContext) return;
         
-        document.body.appendChild(notification);
-        
-        // Animate in
-        setTimeout(() => {
-            notification.classList.add('show');
-        }, 100);
-        
-        // Remove after delay
-        setTimeout(() => {
-            notification.classList.remove('show');
-            setTimeout(() => {
-                if (notification.parentNode) {
-                    notification.remove();
+        try {
+            this.audioContext = new (window.AudioContext || window.webkitAudioContext)();
+            
+            const resumeAudio = () => {
+                if (this.audioContext.state === 'suspended') {
+                    this.audioContext.resume();
                 }
-            }, 300);
-        }, 3000);
+                document.removeEventListener('click', resumeAudio);
+                document.removeEventListener('keydown', resumeAudio);
+            };
+            
+            document.addEventListener('click', resumeAudio);
+            document.addEventListener('keydown', resumeAudio);
+            
+            console.log('🔊 Audio context initialized');
+        } catch (error) {
+            console.warn('⚠️ Failed to setup audio context:', error);
+        }
     }
-    
-    updateThemeComponents(themeName) {
-        // Update charts and graphs if they exist
-        this.updateCharts(themeName);
+
+    playThemeChangeSound() {
+        if (!this.audioContext || this.audioContext.state !== 'running') return;
         
-        // Update code syntax highlighting
-        this.updateCodeHighlighting(themeName);
-        
-        // Update any custom components
-        this.updateCustomComponents(themeName);
+        try {
+            const oscillator = this.audioContext.createOscillator();
+            const gainNode = this.audioContext.createGain();
+            const filter = this.audioContext.createBiquadFilter();
+            
+            oscillator.connect(filter);
+            filter.connect(gainNode);
+            gainNode.connect(this.audioContext.destination);
+            
+            filter.type = 'lowpass';
+            filter.frequency.setValueAtTime(800, this.audioContext.currentTime);
+            filter.frequency.exponentialRampToValueAtTime(200, this.audioContext.currentTime + 0.5);
+            
+            oscillator.frequency.setValueAtTime(440, this.audioContext.currentTime);
+            oscillator.frequency.exponentialRampToValueAtTime(880, this.audioContext.currentTime + 0.3);
+            oscillator.type = 'sine';
+            
+            gainNode.gain.setValueAtTime(0, this.audioContext.currentTime);
+            gainNode.gain.linearRampToValueAtTime(0.3, this.audioContext.currentTime + 0.1);
+            gainNode.gain.exponentialRampToValueAtTime(0.01, this.audioContext.currentTime + 0.5);
+            
+            oscillator.start(this.audioContext.currentTime);
+            oscillator.stop(this.audioContext.currentTime + 0.5);
+            
+        } catch (error) {
+            console.warn('⚠️ Failed to play theme change sound:', error);
+        }
     }
-    
-    updateCharts(themeName) {
-        // Update Chart.js charts if they exist
-        if (window.Chart && window.Chart.instances) {
-            Object.values(window.Chart.instances).forEach(chart => {
-                const colors = this.themeColors[themeName];
-                chart.data.datasets.forEach(dataset => {
-                    if (dataset.backgroundColor === 'rgba(0, 123, 255, 0.2)') {
-                        dataset.backgroundColor = this.hexToRgba(colors.primary, 0.2);
-                        dataset.borderColor = colors.primary;
+
+    playDeviceSelectionSound() {
+        if (!this.audioContext || this.audioContext.state !== 'running') return;
+        
+        try {
+            const oscillator = this.audioContext.createOscillator();
+            const gainNode = this.audioContext.createGain();
+            const filter = this.audioContext.createBiquadFilter();
+            
+            oscillator.connect(filter);
+            filter.connect(gainNode);
+            gainNode.connect(this.audioContext.destination);
+            
+            filter.type = 'highpass';
+            filter.frequency.setValueAtTime(200, this.audioContext.currentTime);
+            filter.frequency.exponentialRampToValueAtTime(800, this.audioContext.currentTime + 0.3);
+            
+            oscillator.frequency.setValueAtTime(220, this.audioContext.currentTime);
+            oscillator.frequency.exponentialRampToValueAtTime(440, this.audioContext.currentTime + 0.2);
+            oscillator.type = 'square';
+            
+            gainNode.gain.setValueAtTime(0, this.audioContext.currentTime);
+            gainNode.gain.linearRampToValueAtTime(0.2, this.audioContext.currentTime + 0.05);
+            gainNode.gain.exponentialRampToValueAtTime(0.01, this.audioContext.currentTime + 0.3);
+            
+            oscillator.start(this.audioContext.currentTime);
+            oscillator.stop(this.audioContext.currentTime + 0.3);
+            
+        } catch (error) {
+            console.warn('⚠️ Failed to play device selection sound:', error);
+        }
+    }
+
+    setupEventListeners() {
+        try {
+            // Theme switching
+            document.addEventListener('click', (e) => {
+                const themeButton = e.target.closest('[data-theme]');
+                if (themeButton) {
+                    e.preventDefault();
+                    const theme = themeButton.dataset.theme;
+                    this.switchTheme(theme);
+                }
+            });
+
+            // Device selection
+            document.addEventListener('click', (e) => {
+                const deviceOption = e.target.closest('.device-option');
+                if (deviceOption) {
+                    e.preventDefault();
+                    const device = deviceOption.dataset.device;
+                    this.selectDevice(device);
+                }
+            });
+
+            // Keyboard shortcuts
+            if (this.features.shortcuts) {
+                document.addEventListener('keydown', (e) => {
+                    if (e.ctrlKey || e.metaKey) {
+                        e.preventDefault();
+                        switch(e.key) {
+                            case '1': this.switchTheme('light'); break;
+                            case '2': this.switchTheme('dark'); break;
+                            case '3': this.switchTheme('cs2'); break;
+                            case '4': this.switchTheme('premium'); break;
+                            case 'd': this.selectDevice('desktop'); break;
+                            case 't': this.selectDevice('tablet'); break;
+                            case 'm': this.selectDevice('mobile'); break;
+                            case 'h': this.toggleHighContrast(); break;
+                            case 'r': this.toggleReducedMotion(); break;
+                            case 'f': this.toggleFontSize(); break;
+                            case 'p': this.toggleParticles(); break;
+                            case 'a': this.toggleAudio(); break;
+                        }
                     }
                 });
-                chart.update();
-            });
-        }
-    }
-    
-    updateCodeHighlighting(themeName) {
-        // Update Prism.js if it exists
-        if (window.Prism) {
-            Prism.highlightAll();
-        }
-        
-        // Update custom code blocks
-        const codeBlocks = document.querySelectorAll('pre code');
-        codeBlocks.forEach(block => {
-            block.style.background = `var(--code-bg)`;
-            block.style.color = `var(--code-text)`;
-        });
-    }
-    
-    updateCustomComponents(themeName) {
-        // Update any custom components that need theme awareness
-        const customComponents = document.querySelectorAll('[data-theme-aware]');
-        customComponents.forEach(component => {
-            component.setAttribute('data-current-theme', themeName);
-        });
-    }
-    
-    hexToRgba(hex, alpha) {
-        const r = parseInt(hex.slice(1, 3), 16);
-        const g = parseInt(hex.slice(3, 5), 16);
-        const b = parseInt(hex.slice(5, 7), 16);
-        return `rgba(${r}, ${g}, ${b}, ${alpha})`;
-    }
-    
-    showThemePreview() {
-        const modal = document.getElementById('theme-preview-modal');
-        if (modal) {
-            modal.style.display = 'block';
-        }
-    }
-    
-    openCustomization() {
-        // Open theme customization panel
-        console.log('Opening theme customization...');
-        // This can be expanded later
-    }
-    
-    previewTheme(themeName) {
-        // Temporarily apply theme for preview
-        const originalTheme = this.currentTheme;
-        document.body.setAttribute('data-theme', themeName);
-        
-        // Revert after 3 seconds
-        setTimeout(() => {
-            document.body.setAttribute('data-theme', originalTheme);
-        }, 3000);
-    }
-    
-    hideThemePreview() {
-        const modal = document.querySelector('.theme-preview-modal');
-        if (modal) {
-            modal.classList.remove('show');
-        }
-    }
-    
-    toggleCustomization() {
-        const panel = document.querySelector('.theme-customization-panel');
-        if (panel) {
-            panel.classList.toggle('show');
-            
-            if (panel.classList.contains('show')) {
-                this.loadCustomizationValues();
             }
-        }
-    }
-    
-    loadCustomizationValues() {
-        const saved = this.getCustomization();
-        
-        // Load saved values
-        const primaryColor = document.getElementById('custom-primary');
-        if (primaryColor) {
-            primaryColor.value = saved.primaryColor || this.themeColors[this.currentTheme].primary;
-        }
-        
-        const opacitySlider = document.getElementById('custom-opacity');
-        if (opacitySlider) {
-            opacitySlider.value = saved.opacity || 1;
-        }
-        
-        const radiusSlider = document.getElementById('custom-radius');
-        if (radiusSlider) {
-            radiusSlider.value = saved.borderRadius || 8;
-        }
-        
-        const animationSelect = document.getElementById('custom-animation');
-        if (animationSelect) {
-            animationSelect.value = saved.animationSpeed || 'normal';
-        }
-    }
-    
-    updateCustomColor(type, value) {
-        document.documentElement.style.setProperty(`--custom-${type}`, value);
-    }
-    
-    updateCustomOpacity(value) {
-        document.documentElement.style.setProperty('--custom-opacity', value);
-    }
-    
-    updateCustomRadius(value) {
-        document.documentElement.style.setProperty('--custom-border-radius', `${value}px`);
-    }
-    
-    updateCustomAnimation(value) {
-        const speeds = {
-            fast: '150ms',
-            normal: '250ms',
-            slow: '400ms'
-        };
-        document.documentElement.style.setProperty('--custom-transition', speeds[value]);
-    }
-    
-    saveCustomization() {
-        const customization = {
-            primaryColor: document.getElementById('custom-primary')?.value,
-            opacity: document.getElementById('custom-opacity')?.value,
-            borderRadius: document.getElementById('custom-radius')?.value,
-            animationSpeed: document.getElementById('custom-animation')?.value
-        };
-        
-        localStorage.setItem('affinity_customization', JSON.stringify(customization));
-        
-        // Show success message
-        this.showCustomizationNotification('Customization saved!', 'success');
-        
-        // Hide panel
-        this.toggleCustomization();
-    }
-    
-    resetCustomization() {
-        // Reset to default values
-        this.updateCustomColor('primary', this.themeColors[this.currentTheme].primary);
-        this.updateCustomOpacity(1);
-        this.updateCustomRadius(8);
-        this.updateCustomAnimation('normal');
-        
-        // Clear saved customization
-        localStorage.removeItem('affinity_customization');
-        
-        // Show reset message
-        this.showCustomizationNotification('Customization reset to default', 'info');
-        
-        // Reload values
-        this.loadCustomizationValues();
-    }
-    
-    getCustomization() {
-        const saved = localStorage.getItem('affinity_customization');
-        return saved ? JSON.parse(saved) : {};
-    }
-    
-    showCustomizationNotification(message, type = 'info') {
-        // Create notification
-        const notification = document.createElement('div');
-        notification.className = `customization-notification ${type}`;
-        notification.textContent = message;
-        
-        document.body.appendChild(notification);
-        
-        // Show and hide
-        setTimeout(() => notification.classList.add('show'), 100);
-        setTimeout(() => {
-            notification.classList.remove('show');
-            setTimeout(() => notification.remove(), 300);
-        }, 3000);
-    }
-    
-    onThemeChange(themeName) {
-        // Handle any additional theme change logic
-        console.log(`Theme changed to: ${themeName}`);
-        
-        // Update any external components
-        this.updateExternalComponents(themeName);
-    }
-    
-    updateExternalComponents(themeName) {
-        // Update any external libraries or components
-        // This can be extended based on what's used in the forum
-        
-        // Example: Update TinyMCE if it exists
-        if (window.tinymce) {
-            window.tinymce.editors.forEach(editor => {
-                editor.theme.resizeTo(editor.getContainer().offsetWidth, editor.getContainer().offsetHeight);
+            
+            // Performance monitoring
+            window.addEventListener('beforeunload', () => {
+                this.savePerformanceMetrics();
             });
+            
+            console.log('🎯 Event listeners configured');
+        } catch (error) {
+            console.error('❌ Failed to setup event listeners:', error);
+        }
+    }
+
+    setupPerformanceMonitoring() {
+        console.log('📊 Performance monitoring initialized');
+    }
+
+    switchTheme(themeName) {
+        if (this.currentTheme === themeName) return;
+        
+        try {
+            const startTime = performance.now();
+            
+            this.currentTheme = themeName;
+            localStorage.setItem('selectedTheme', themeName);
+            
+            this.applyCurrentTheme();
+            this.updateParticleColors();
+            this.playThemeChangeSound();
+            
+            const switchTime = performance.now() - startTime;
+            this.performanceMetrics.lastThemeSwitch = switchTime;
+            this.performanceMetrics.themeSwitchCount++;
+            
+            this.showNotification(`Theme switched to ${themeName}!`, 'success');
+            
+            document.dispatchEvent(new CustomEvent('themeChanged', {
+                detail: { theme: themeName, switchTime }
+            }));
+            
+            console.log(`🎨 Theme switched to ${themeName} in ${switchTime.toFixed(2)}ms`);
+        } catch (error) {
+            console.error('❌ Failed to switch theme:', error);
+            this.showNotification('Failed to switch theme!', 'error');
+        }
+    }
+
+    updateParticleColors() {
+        if (!this.particleSystem || !this.particleSystem.particles) return;
+        
+        try {
+            this.particleSystem.particles.forEach(particle => {
+                particle.color = this.getRandomThemeColor();
+            });
+            
+            console.log('✨ Particle colors updated for new theme');
+        } catch (error) {
+            console.warn('⚠️ Failed to update particle colors:', error);
+        }
+    }
+
+    selectDevice(device) {
+        console.log('📱 Selecting device:', device);
+        this.currentDevice = device;
+        this.applyCurrentTheme();
+        localStorage.setItem('selectedDevice', device);
+        localStorage.setItem('hasSelectedDevice', 'true');
+        
+        const overlay = document.getElementById('deviceOverlay');
+        if (overlay) {
+            overlay.style.display = 'none';
+            console.log('✅ Device overlay hidden after selection');
         }
         
-        // Example: Update Summernote if it exists
-        if (window.summernote) {
-            // Summernote specific updates
+        this.updateParticleSystemForDevice(device);
+        this.playDeviceSelectionSound();
+        
+        console.log('✅ Device selected successfully');
+    }
+
+    updateParticleSystemForDevice(device) {
+        if (!this.particleSystem) return;
+        
+        try {
+            switch (device) {
+                case 'mobile':
+                    this.particleSystem.particleCount = 20;
+                    this.particleSystem.maxParticles = 40;
+                    break;
+                case 'tablet':
+                    this.particleSystem.particleCount = 35;
+                    this.particleSystem.maxParticles = 70;
+                    break;
+                case 'desktop':
+                default:
+                    this.particleSystem.particleCount = 50;
+                    this.particleSystem.maxParticles = 100;
+                    break;
+            }
+            
+            console.log(`📱 Particle system optimized for ${device}`);
+        } catch (error) {
+            console.warn('⚠️ Failed to update particle system for device:', error);
+        }
+    }
+
+    applyCurrentTheme() {
+        try {
+            document.body.classList.remove('theme-light', 'theme-dark', 'theme-cs2', 'theme-premium');
+            document.body.classList.add(`theme-${this.currentTheme}`);
+            document.body.setAttribute('data-theme', this.currentTheme);
+            document.body.setAttribute('data-device', this.currentDevice);
+            
+            this.updateCSSVariables();
+            
+            if (this.particleSystem) {
+                this.updateParticleColors();
+            }
+            
+            console.log(`🎨 Applied theme: ${this.currentTheme}, device: ${this.currentDevice}`);
+        } catch (error) {
+            console.error('❌ Failed to apply current theme:', error);
+        }
+    }
+
+    updateCSSVariables() {
+        try {
+            const root = document.documentElement;
+            
+            const themeColors = this.getThemeColors(this.currentTheme);
+            Object.entries(themeColors).forEach(([key, value]) => {
+                root.style.setProperty(`--${key}`, value);
+            });
+            
+            const deviceVars = this.getDeviceVariables(this.currentDevice);
+            Object.entries(deviceVars).forEach(([key, value]) => {
+                root.style.setProperty(`--${key}`, value);
+            });
+            
+        } catch (error) {
+            console.warn('⚠️ Failed to update CSS variables:', error);
+        }
+    }
+
+    getThemeColors(theme) {
+        const themes = {
+            light: {
+                'primary-color': '#007bff',
+                'secondary-color': '#6c757d',
+                'accent-color': '#28a745',
+                'text-color': '#212529',
+                'bg-color': '#ffffff',
+                'border-color': '#dee2e6'
+            },
+            dark: {
+                'primary-color': '#0d6efd',
+                'secondary-color': '#6c757d',
+                'accent-color': '#198754',
+                'text-color': '#f8f9fa',
+                'bg-color': '#212529',
+                'border-color': '#495057'
+            },
+            cs2: {
+                'primary-color': '#ff6b35',
+                'secondary-color': '#2c3e50',
+                'accent-color': '#e74c3c',
+                'text-color': '#ecf0f1',
+                'bg-color': '#1a1a1a',
+                'border-color': '#34495e'
+            },
+            premium: {
+                'primary-color': '#d4af37',
+                'secondary-color': '#2c3e50',
+                'accent-color': '#e74c3c',
+                'text-color': '#f8f9fa',
+                'bg-color': '#1a1a1a',
+                'border-color': '#d4af37'
+            }
+        };
+        
+        return themes[theme] || themes.light;
+    }
+
+    getDeviceVariables(device) {
+        const devices = {
+            desktop: {
+                'container-width': '1200px',
+                'font-size': '16px',
+                'spacing': '1.5rem',
+                'border-radius': '10px'
+            },
+            tablet: {
+                'container-width': '768px',
+                'font-size': '14px',
+                'spacing': '1rem',
+                'border-radius': '8px'
+            },
+            mobile: {
+                'container-width': '100%',
+                'font-size': '12px',
+                'spacing': '0.75rem',
+                'border-radius': '6px'
+            }
+        };
+        
+        return devices[device] || devices.desktop;
+    }
+
+    startParticleAnimation() {
+        if (!this.particleSystem || !this.particleSystem.isActive) return;
+        
+        const animate = () => {
+            if (!this.particleSystem.isActive) return;
+            
+            const { ctx, particles, maxParticles } = this.particleSystem;
+            
+            ctx.fillStyle = 'rgba(0, 0, 0, 0.05)';
+            ctx.fillRect(0, 0, ctx.canvas.width, ctx.canvas.height);
+            
+            for (let i = particles.length - 1; i >= 0; i--) {
+                const particle = particles[i];
+                
+                particle.x += particle.vx;
+                particle.y += particle.vy;
+                particle.pulse += particle.pulseSpeed;
+                particle.life--;
+                
+                if (particle.life <= 0 || 
+                    particle.x < 0 || particle.x > ctx.canvas.width ||
+                    particle.y < 0 || particle.y > ctx.canvas.height) {
+                    particles.splice(i, 1);
+                    continue;
+                }
+                
+                ctx.save();
+                ctx.globalAlpha = particle.opacity * (particle.life / particle.maxLife);
+                
+                const pulseScale = 1 + Math.sin(particle.pulse) * 0.3;
+                const size = particle.size * pulseScale;
+                
+                const gradient = ctx.createRadialGradient(
+                    particle.x, particle.y, 0,
+                    particle.x, particle.y, size
+                );
+                gradient.addColorStop(0, particle.color);
+                gradient.addColorStop(1, 'transparent');
+                
+                ctx.fillStyle = gradient;
+                ctx.beginPath();
+                ctx.arc(particle.x, particle.y, size, 0, Math.PI * 2);
+                ctx.fill();
+                
+                ctx.shadowColor = particle.color;
+                ctx.shadowBlur = size * 2;
+                ctx.beginPath();
+                ctx.arc(particle.x, particle.y, size * 0.5, 0, Math.PI * 2);
+                ctx.fill();
+                
+                ctx.restore();
+            }
+            
+            if (particles.length < maxParticles && Math.random() < 0.1) {
+                this.addNewParticle();
+            }
+            
+            this.particleSystem.animationId = requestAnimationFrame(animate);
+        };
+        
+        animate();
+    }
+
+    addNewParticle() {
+        if (!this.particleSystem) return;
+        
+        const { ctx, particles } = this.particleSystem;
+        
+        particles.push({
+            x: Math.random() * ctx.canvas.width,
+            y: Math.random() * ctx.canvas.height,
+            vx: (Math.random() - 0.5) * 2,
+            vy: (Math.random() - 0.5) * 2,
+            size: Math.random() * 3 + 1,
+            opacity: Math.random() * 0.5 + 0.3,
+            color: this.getRandomThemeColor(),
+            pulse: Math.random() * Math.PI * 2,
+            pulseSpeed: Math.random() * 0.02 + 0.01,
+            life: Math.random() * 100 + 50,
+            maxLife: 150
+        });
+    }
+
+    getRandomThemeColor() {
+        const colors = ['#007bff', '#28a745', '#ffc107', '#dc3545', '#6f42c1', '#fd7e14'];
+        return colors[Math.floor(Math.random() * colors.length)];
+    }
+
+    showNotification(message, type = 'info', duration = 5000) {
+        if (!this.features.notifications) return;
+        
+        try {
+            const notification = document.createElement('div');
+            notification.className = `notification notification-${type}`;
+            notification.innerHTML = `
+                <div class="notification-content">
+                    <div class="notification-icon">${this.getNotificationIcon(type)}</div>
+                    <div class="notification-message">${message}</div>
+                    <button class="notification-close" onclick="this.parentElement.parentElement.remove()">×</button>
+                </div>
+                <div class="notification-progress"></div>
+            `;
+            
+            const container = document.getElementById('notificationContainer');
+            if (container) {
+                container.appendChild(notification);
+                
+                if (container.children.length > 5) {
+                    container.removeChild(container.firstChild);
+                }
+                
+                if (duration > 0) {
+                    setTimeout(() => {
+                        if (notification.parentNode) {
+                            notification.remove();
+                        }
+                    }, duration);
+                }
+                
+                const progress = notification.querySelector('.notification-progress');
+                if (progress && duration > 0) {
+                    progress.style.transition = `width ${duration}ms linear`;
+                    setTimeout(() => {
+                        progress.style.width = '0%';
+                    }, 100);
+                }
+                
+                this.performanceMetrics.totalNotifications++;
+                
+                setTimeout(() => {
+                    notification.style.transform = 'translateX(0)';
+                    notification.style.opacity = '1';
+                }, 100);
+            }
+        } catch (error) {
+            console.error('❌ Failed to show notification:', error);
+        }
+    }
+
+    getNotificationIcon(type) {
+        const icons = {
+            success: '✅',
+            error: '❌',
+            warning: '⚠️',
+            info: 'ℹ️'
+        };
+        return icons[type] || icons.info;
+    }
+
+    // Utility methods
+    getCurrentTheme() {
+        return this.currentTheme;
+    }
+    
+    getCurrentDevice() {
+        return this.currentDevice;
+    }
+    
+    isThemeActive(themeName) {
+        return this.currentTheme === themeName;
+    }
+    
+    isDeviceActive(deviceName) {
+        return this.currentDevice === deviceName;
+    }
+
+    // Accessibility toggles
+    toggleHighContrast() {
+        this.accessibility.highContrast = !this.accessibility.highContrast;
+        localStorage.setItem('highContrast', this.accessibility.highContrast);
+        
+        if (this.accessibility.highContrast) {
+            document.body.classList.add('high-contrast');
+            this.showNotification('High contrast mode enabled!', 'success');
+        } else {
+            document.body.classList.remove('high-contrast');
+            this.showNotification('High contrast mode disabled!', 'info');
+        }
+    }
+
+    toggleReducedMotion() {
+        this.accessibility.reducedMotion = !this.accessibility.reducedMotion;
+        localStorage.setItem('reducedMotion', this.accessibility.reducedMotion);
+        
+        if (this.accessibility.reducedMotion) {
+            document.body.classList.add('reduced-motion');
+            this.showNotification('Reduced motion enabled!', 'success');
+        } else {
+            document.body.classList.remove('reduced-motion');
+            this.showNotification('Reduced motion disabled!', 'info');
+        }
+    }
+
+    toggleFontSize() {
+        const sizes = [12, 14, 16, 18, 20, 22, 24];
+        const currentIndex = sizes.indexOf(this.accessibility.fontSize);
+        const nextIndex = (currentIndex + 1) % sizes.length;
+        this.accessibility.fontSize = sizes[nextIndex];
+        
+        localStorage.setItem('fontSize', this.accessibility.fontSize);
+        document.documentElement.style.fontSize = this.accessibility.fontSize + 'px';
+        
+        this.showNotification(`Font size: ${this.accessibility.fontSize}px`, 'info');
+    }
+
+    toggleParticles() {
+        this.features.particles = !this.features.particles;
+        localStorage.setItem('particles', this.features.particles);
+        
+        if (this.features.particles) {
+            this.setupParticleSystem();
+            this.showNotification('Particle effects enabled!', 'success');
+        } else {
+            if (this.particleSystem && this.particleSystem.canvas) {
+                this.particleSystem.canvas.remove();
+                this.particleSystem = null;
+            }
+            this.showNotification('Particle effects disabled!', 'warning');
+        }
+    }
+
+    toggleAudio() {
+        this.features.audio = !this.features.audio;
+        localStorage.setItem('audio', this.features.audio);
+        
+        if (this.features.audio) {
+            this.setupAudioContext();
+            this.showNotification('Audio feedback enabled!', 'success');
+        } else {
+            if (this.audioContext) {
+                this.audioContext.close();
+                this.audioContext = null;
+            }
+            this.showNotification('Audio feedback disabled!', 'warning');
+        }
+    }
+
+    savePerformanceMetrics() {
+        try {
+            localStorage.setItem('themeSwitchCount', this.performanceMetrics.themeSwitchCount);
+            localStorage.setItem('deviceSwitchCount', this.performanceMetrics.deviceSwitchCount);
+            localStorage.setItem('lastThemeSwitch', this.performanceMetrics.lastThemeSwitch);
+            localStorage.setItem('lastDeviceSwitch', this.performanceMetrics.lastDeviceSwitch);
+            localStorage.setItem('totalNotifications', this.performanceMetrics.totalNotifications);
+            console.log('📈 Performance metrics saved');
+        } catch (error) {
+            console.warn('⚠️ Failed to save performance metrics:', error);
+        }
+    }
+
+    getPerformanceReport() {
+        return {
+            ...this.performanceMetrics,
+            currentTheme: this.currentTheme,
+            currentDevice: this.currentDevice,
+            features: this.features,
+            accessibility: this.accessibility,
+            timestamp: Date.now()
+        };
+    }
+
+    optimizePerformance() {
+        try {
+            // Reduce particle count on low-end devices
+            if (navigator.hardwareConcurrency && navigator.hardwareConcurrency < 4) {
+                if (this.particleSystem) {
+                    this.particleSystem.particleCount = 25;
+                    this.particleSystem.maxParticles = 50;
+                }
+            }
+            
+            // Disable animations on low-end devices
+            if (navigator.deviceMemory && navigator.deviceMemory < 4) {
+                this.features.animations = false;
+                localStorage.setItem('animations', 'false');
+                document.body.classList.add('reduced-motion');
+            }
+            
+            // Optimize for mobile
+            if (window.innerWidth < 768) {
+                if (this.particleSystem) {
+                    this.particleSystem.particleCount = 20;
+                }
+                this.features.animations = false;
+                localStorage.setItem('animations', 'false');
+            }
+            
+            // Reduce motion if user prefers
+            if (window.matchMedia('(prefers-reduced-motion: reduce)').matches) {
+                this.accessibility.reducedMotion = true;
+                localStorage.setItem('reducedMotion', 'true');
+                document.body.classList.add('reduced-motion');
+            }
+            
+            console.log('🚀 Performance optimized');
+        } catch (error) {
+            console.warn('⚠️ Failed to optimize performance:', error);
+        }
+    }
+
+    getAvailableThemes() {
+        return [
+            { id: 'light', name: 'Light Theme', icon: '☀️', description: 'Clean and modern light interface' },
+            { id: 'dark', name: 'Dark Theme', icon: '🌙', description: 'Sleek and elegant dark interface' },
+            { id: 'cs2', name: 'CS2 Gaming', icon: '🎮', description: 'Gaming-inspired Counter-Strike 2 style' },
+            { id: 'premium', name: 'Premium Luxury', icon: '💎', description: 'Exclusive luxury premium interface' }
+        ];
+    }
+
+    getAvailableDevices() {
+        return [
+            { id: 'desktop', name: 'Desktop', icon: '🖥️', description: 'Full desktop experience' },
+            { id: 'tablet', name: 'Tablet', icon: '📱', description: 'Optimized for tablet devices' },
+            { id: 'mobile', name: 'Mobile', icon: '📱', description: 'Mobile-optimized interface' }
+        ];
+    }
+
+    resetToDefaults() {
+        try {
+            localStorage.clear();
+            
+            this.currentTheme = 'light';
+            this.currentDevice = 'desktop';
+            this.accessibility = {
+                highContrast: false,
+                reducedMotion: false,
+                fontSize: 16,
+                screenReader: false
+            };
+            this.features = {
+                particles: true,
+                audio: true,
+                animations: true,
+                notifications: true,
+                shortcuts: true
+            };
+            
+            this.applyCurrentTheme();
+            this.setupAccessibility();
+            
+            this.showNotification('Settings reset to defaults!', 'success');
+        } catch (error) {
+            console.error('❌ Failed to reset to defaults:', error);
+        }
+    }
+
+    exportSettings() {
+        try {
+            const settings = {
+                theme: this.currentTheme,
+                device: this.currentDevice,
+                accessibility: this.accessibility,
+                features: this.features,
+                performanceMetrics: this.performanceMetrics,
+                timestamp: Date.now()
+            };
+            
+            const dataStr = JSON.stringify(settings, null, 2);
+            const dataBlob = new Blob([dataStr], { type: 'application/json' });
+            
+            const link = document.createElement('a');
+            link.href = URL.createObjectURL(dataBlob);
+            link.download = `affinity-theme-settings-${Date.now()}.json`;
+            link.click();
+            
+            this.showNotification('Settings exported successfully!', 'success');
+            return true;
+        } catch (error) {
+            console.error('❌ Failed to export settings:', error);
+            return false;
+        }
+    }
+
+    importSettings(file) {
+        try {
+            const reader = new FileReader();
+            reader.onload = (e) => {
+                const settings = JSON.parse(e.target.result);
+                
+                if (settings.theme && ['light', 'dark', 'cs2', 'premium'].includes(settings.theme)) {
+                    this.currentTheme = settings.theme;
+                    localStorage.setItem('selectedTheme', settings.theme);
+                }
+                
+                if (settings.device && ['desktop', 'tablet', 'mobile'].includes(settings.device)) {
+                    this.currentDevice = settings.device;
+                    localStorage.setItem('selectedDevice', settings.device);
+                }
+                
+                this.applyCurrentTheme();
+                this.showNotification('Settings imported successfully!', 'success');
+            };
+            
+            reader.readAsText(file);
+            return true;
+        } catch (error) {
+            console.error('❌ Failed to import settings:', error);
+            return false;
         }
     }
 }
 
-// Initialize theme manager when DOM is ready
-document.addEventListener('DOMContentLoaded', function() {
-    console.log('🎨 Initializing Theme Manager...');
-    window.themeManager = new ThemeManager();
-    console.log('✅ Theme Manager initialized successfully!');
+// Create global instance
+window.themeManager = new ThemeManager();
+console.log('✅ Global themeManager instance created:', window.themeManager);
+
+// Global functions for HTML onclick
+window.switchTheme = (themeName) => {
+    if (window.themeManager) {
+        window.themeManager.switchTheme(themeName);
+    }
+};
+
+window.selectDevice = (device) => {
+    if (window.themeManager) {
+        window.themeManager.selectDevice(device);
+    }
+};
+
+window.toggleHighContrast = () => {
+    if (window.themeManager) {
+        window.themeManager.toggleHighContrast();
+    }
+};
+
+window.toggleReducedMotion = () => {
+    if (window.themeManager) {
+        window.themeManager.toggleReducedMotion();
+    }
+};
+
+window.toggleFontSize = () => {
+    if (window.themeManager) {
+        window.themeManager.toggleFontSize();
+    }
+};
+
+window.toggleParticles = () => {
+    if (window.themeManager) {
+        window.themeManager.toggleParticles();
+    }
+};
+
+window.toggleAudio = () => {
+    if (window.themeManager) {
+        window.themeManager.toggleAudio();
+    }
+};
+
+window.exportSettings = () => {
+    if (window.themeManager) {
+        window.themeManager.exportSettings();
+    }
+};
+
+window.importSettings = (file) => {
+    if (window.themeManager) {
+        window.themeManager.importSettings(file);
+    }
+};
+
+window.resetToDefaults = () => {
+    if (window.themeManager) {
+        window.themeManager.resetToDefaults();
+    }
+};
+
+// Initialize when DOM is ready
+document.addEventListener('DOMContentLoaded', () => {
+    console.log('🚀 DOM ready - Theme manager initialized');
+    
+    if (window.themeManager) {
+        window.themeManager.optimizePerformance();
+    }
 });
 
-// Export for module systems
-if (typeof module !== 'undefined' && module.exports) {
-    module.exports = ThemeManager;
-}
+console.log('🎨 Unified Theme Manager loaded successfully!');
